Add support for multiple gatling sims in a jenkins run
This commit tweaks the logic used by the GatlingPublisher so that
it will find all of the Gatling report directories that were
created after the start time of the jenkins build, and archive
all of them.

It also tweaks the trend graphs to show series data for all
of the different gatling reports for each build, rather than
just a single one.

Finally, it changes the behavior of what happens when you click
on the "Gatling" link in the sidebar of an individual jenkins
build so that rather than immediately rendering the report,
it instead provides a list of links for all of the reports that
were found.  These links then render the report as it was before.