(maint) Make graph data points clickable
Prior to this commit, the data points within the gatling graphs only
responded to a mouse hovering over an individual data point with a tool
tip. This commit changes that by allowing the user to click on a data
point, which will then open a new window and take them to the specific
build the data point is related to.